chore: rearrange folder structure and improve code
